+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.2.0
+
+* added transparent folder/README handling
+* added {{paginate}} theme macro
+* fixed unnecessary markup preprocessing in editor
+* added version/repo fields to wiki class
+* fixed missing dot-files and file extensions in release files (.zip, .tar.gz)
+
 # v0.1.0 - first public release
 
 * initial version